Enable hermetic builds for sbom-utility-scripts
Enabling hermetic builds insulates the build process from the network.
Doing that improves security by preventing outside interference.

In order to build hermetically, dependencies needed to build the
container image have to be prefetched.

The prefetching for sbom-utility-scripts was set up in previous commits,
so the hermetic builds should now be enabled.

Signed-off-by: Tomáš Nevrlka <tnevrlka@redhat.com>